addressed review feedback comments  - refactored code  - fixed the potential recursive loop execution  - replaced string with dict to avoid string comparisons  - changed code to avoid complexity level to number of digits mapping
